Adds Eureka and Zuul.
Adds Eureka service discovery. All the micro services now
can be seen on localhost:8761.
Also adds Zuul proxy server. It is a gateway for upcoming traffic,
which will be like a facade for all services on the backend.
I created routing for loader and data so all the traffic will go through gateway.